Move upgrade_guestos to guestos_upgrade.rs
